pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "madness"
version = "0.1.3"
description = "Personal Website"
authors = [
  {name = 'Aidan Courtney', email = 'aidanfc97@gmail.com'},
]
license = {text = 'MIT License'}
readme = 'README.md'
requires-python = ">=3.12"
dependencies = [
  'fastapi',
  'typer',
  'uvicorn',
  'rich',
  'pyjwt',
  'argon2-cffi',
  'python-multipart',
  'aiosqlite',
  'websockets',
  'types-psutil',
]

[project.optional-dependencies]
dev = ["black", "isort", "ruff", "pytest", "pytest-watch"]

[project.scripts]
madness = "server.__main__:cli"

[tool.setuptools]
packages = ['server']

[tool.ruff]
lint.ignore = ['E501']
target-version = 'py312'

[tool.pyright]
pythonVersion = '3.12'
